vstreamclient: move to new dir
