+++ go/vt/vstreamclient/chunk_test.go
@@ -3,7 +3,7 @@
 // Copyright 2024 The Go Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
-package vstreamer
+package vstreamclient
 
 import (
 	"reflect"
