+++ go/vt/vstreamclient/run.go
@@ -1,4 +1,4 @@
-package vstreamer
+package vstreamclient
 
 import (
 	"context"
